+++ NEWS.md
@@ -1,3 +1,9 @@
+# quickmatch 0.2.0
+
+  * Makes defaults and error messages friendlier with discrete covariates
+  * Changes so `matching_weights` returns zero for unassigned units.
+
+
 # quickmatch 0.1.3
 
   * Adds function for covariate averages.